requirements.txt
streamlit
langchain
google-generativeai
python-dotenv
PyPDF2
faiss-cpu
langchain_google_genai
langchain-community